Refactor gql reservation tests to separate files/modules
TILA-1062
